refactor: extract begin_transaction into core module
