+++ src/core.rs
@@ -1,8 +1,9 @@
-use crate::database::{DatastoreStorage, EntityWithMetadata, KeyStruct};
+use crate::database::{DatastoreStorage, EntityWithMetadata, KeyStruct, TransactionState};
 use crate::google::datastore::v1::{
-    CommitRequest, CommitResponse, Entity, EntityResult, ExecutionStats, ExplainMetrics,
-    LookupRequest, LookupResponse, MutationResult, PlanSummary, RunQueryRequest, RunQueryResponse,
-    commit_request::TransactionSelector, mutation::Operation,
+    BeginTransactionRequest, BeginTransactionResponse, CommitRequest, CommitResponse, Entity,
+    EntityResult, ExecutionStats, ExplainMetrics, LookupRequest, LookupResponse, MutationResult,
+    PlanSummary, RunQueryRequest, RunQueryResponse, commit_request::TransactionSelector,
+    mutation::Operation,
 };
 use pbjson_types::{Duration, Struct, Value as ValueProps, value::Kind};
 use std::collections::HashMap;
@@ -366,4 +367,66 @@ pub async fn commit(
         index_updates,
         commit_time: Some(commit_time),
     })
+}
+
+pub async fn begin_transaction(
+    storage: &Arc<RwLock<DatastoreStorage>>,
+    req: BeginTransactionRequest,
+) -> Result<BeginTransactionResponse, Status> {
+    tracing::debug!("Received BeginTransactionRequest: {:?}", req);
+    let start = SystemTime::now();
+    let duration_since_epoch = start
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap_or_default();
+
+    let timestamp = pbjson_types::Timestamp {
+        seconds: duration_since_epoch.as_secs() as i64,
+        nanos: duration_since_epoch.as_nanos() as i32,
+    };
+
+    // Generate a unique transaction ID and create transaction state
+    let transaction_id;
+    {
+        let mut storage = storage.write().await;
+
+        // Generate transaction ID using timestamp and current counter
+        let counter = storage
+            .transaction_counter
+            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
+        transaction_id = format!("tx-{}-{}", timestamp.seconds, counter);
+
+        // Check if the transaction is read-only
+        let transaction_options = req.transaction_options.unwrap_or_default();
+        let read_only = if let Some(mode) = transaction_options.mode {
+            match mode {
+                crate::google::datastore::v1::transaction_options::Mode::ReadOnly(_) => true,
+                crate::google::datastore::v1::transaction_options::Mode::ReadWrite(_) => false,
+            }
+        } else {
+            // Default to read-write if no mode is specified
+            false
+        };
+
+        // Create a new transaction state
+        let transaction_state = TransactionState {
+            mutations: Vec::new(),
+            snapshot: HashMap::new(),
+            timestamp,
+            read_only,
+        };
+
+        storage
+            .transactions
+            .insert(transaction_id.clone(), transaction_state);
+    }
+
+    let transaction_bytes = transaction_id.into_bytes();
+    let transaction_response = BeginTransactionResponse {
+        transaction: transaction_bytes,
+    };
+    tracing::debug!(
+        "Began transaction with ID: {:?}",
+        String::from_utf8_lossy(&transaction_response.transaction)
+    );
+    Ok(transaction_response)
 }
